Import Debian changes 1.11-8
trustedqsl (1.11-8) unstable; urgency=low

  * Fix a couple of lintian warnings:
    - call dh_installdeb after calling dh_shlibdeps and dh_makeshlibs.
    - set debhelper compatibility version to 5.
    - update standards version.
    - install and create xpm files for the debian menu.
  * New icons for the desktop menu files, requires imagemagick when
    building the package.
  * Fix all compilation warnings, most of them fixed by changing
    'char *' to 'const char *'.
  * wxNotebookSizer is deprecated, we can now use wxNotebook directly.
  * Build against libwxgtk2.8-dev. tqslcert and tqsl have been ported
    to version 2.8 of libwxgtk.

+++ stationdial.cpp
@@ -383,7 +383,7 @@ TQSLGetStationNameDialog::SelectName(const wxString& name) {
 	namelist->GetSelections(sels);
 	for (int i = 0; i < (int)sels.GetCount(); i++)
 		namelist->Deselect(sels[i]);
-	for (int i = 0; i < namelist->GetCount(); i++) {
+	for (int i = 0; i < (int)namelist->GetCount(); i++) {
 		if (name == *(wxString *)namelist->GetClientData(i)) {
 			namelist->SetSelection(i, TRUE);
 			break;
